Introduce custom_properties in configurations
