+++ src/commands/qbscommands.ts
@@ -5,6 +5,7 @@ export * from './qbscleancommand';
 export * from './qbsdebugproductcommand';
 export * from './qbsdetectprofilescommand';
 export * from './qbsgetbuilddirectory'
+export * from './qbsgetconfigurationparams';
 export * from './qbsgetrunenvironmentcommand';
 export * from './qbsgetselectedproductpath';
 export * from './qbsinstallcommand';
@@ -165,4 +166,7 @@ export async function subscribeCommands(ctx: vscode.ExtensionContext, session: Q
     ctx.subscriptions.push(vscode.commands.registerCommand(QbsCommandKey.GetConfigurationName, async () => {
         return QbsCommand.getConfigurationName(session);
     }));
+    ctx.subscriptions.push(vscode.commands.registerCommand(QbsCommandKey.GetCustomProperty, async (property: string) => {
+        return QbsCommand.getCustomProperty(session, property);
+    }))
 }
