+++ src/datatypes/qbskeys.ts
@@ -2,6 +2,7 @@ export enum QbsDataKey {
     ApiCompatLevel = 'api-compat-level',
     ApiLevel = 'api-level',
     Arguments = 'arguments',
+    AssociatedProfile = 'profile',
     BuildDirectory = 'build-directory',
     BuildRoot = 'build-root',
     BuildSystemFiles = 'build-system-files',
@@ -11,6 +12,7 @@ export enum QbsDataKey {
     CommandEchoMode = 'command-echo-mode',
     ConfigurationName = 'configuration-name',
     Cpp = "cpp",
+    CustomProperties = 'custom_properties',
     DataMode = 'data-mode',
     Description = 'description',
     DisplayName = 'displayName',
@@ -59,7 +61,6 @@ export enum QbsDataKey {
     ProjectFilePath = 'project-file-path',
     ProjectResolved = 'project-resolved',
     Properties = 'properties',
-    AssociatedProfile = 'profile',
     ProtocolError = 'protocol-error',
     Qbs = 'qbs',
     RunEnvironment = 'run-environment',
